+++ README.md
@@ -10,6 +10,6 @@ Run apollo server.
 yarn && yarn start
 ```
 
-Then, you can request the schema for `http://localhost:4000` with some tools like Postman, Insomnia and Apollo Studio.
+Then, you can request the schema for `http://localhost:4000` with some tools like Postman, Insomnia and Apollo Sandbox.
 
-[Apllo Studio](https://studio.apollographql.com/sandbox/explorer/?_gl=1%2A16er0h%2A_ga%2AMTIwMDgyOTQwOS4xNzE3ODAzMTgy%2A_ga_0BGG5V2W2K%2AMTcxNzgwMzE4Mi4xLjEuMTcxNzgwMzc5NC4wLjAuMA..)
+[Apllo Sandbox](https://studio.apollographql.com/sandbox/explorer/?_gl=1%2A16er0h%2A_ga%2AMTIwMDgyOTQwOS4xNzE3ODAzMTgy%2A_ga_0BGG5V2W2K%2AMTcxNzgwMzE4Mi4xLjEuMTcxNzgwMzc5NC4wLjAuMA..)
