+++ src/data.ts
@@ -1,0 +1,31 @@
+export const books = [
+  {
+    title: "The Awakening",
+    author: {
+      name: "Kate Chopin",
+      books: [],
+    },
+  },
+  {
+    title: "City of Glass",
+    author: {
+      name: "Paul Auster",
+      books: [],
+    },
+  },
+];
+
+export const authors = [
+  {
+    name: "Kate Chopin",
+    books: {
+      title: "Kate Chopin",
+    },
+  },
+  {
+    name: "Paul Auster",
+    books: {
+      title: "City of Glass",
+    },
+  },
+];
